+++ earth2studio/data/__init__.py
@@ -15,6 +15,7 @@
 # limitations under the License.
 
 from .base import DataSource  # noqa
+from .wb2 import WB2Climatology
 from .arco import ARCO  # noq
 from .cds import CDS  # noq
 from .gfs import GFS  # noq
@@ -23,4 +24,4 @@ from .ifs import IFS  # noq
 from .rand import Random  # noqa
 from .rx import LandSeaMask, SurfaceGeoPotential, CosineSolarZenith  # noqa
 from .xr import DataArrayFile, DataSetFile  # noqa
-from .utils import fetch_data, prep_data_array
+from .utils import datasource_to_file, fetch_data, prep_data_array